Chapter 8: Add login, logout, temp session and update UI.
The session is lost on logout or after browser is closed.

On a successful login, the header now shows an Account tab with a
dropdown and a Users tab which does nothing right now.

A bit confused about why the @current_user is an instance var in the
SessionsHelper module (and therefore in the ApplicationController
class) which means different instances of controllers could have a
diff copy. Shouldn't that cause a problem?