11.2 Setup account activation email.
Update the mailer, views and preview. Add a test for the account
activation email. Setup the dev/test configs for mailer.

Hook this up with User#create to send out an activation email instead
of automatically logging in the user.

Turn off the part of the user signup test that checks for redirection
to the user profile