13.3.2 Add form in the home page to submit the microposts
When logged in, the home page display the user info and a form to
submit the micro posts, both via separate partials. Since this comes
via the StaticPages controller, populate @microposts there.

When not logged in, it displays the usual static home page.

Make shared/error_messages work with any passed in object and reuse
that in the form above.

Update styling for aside.section.span