13.3.3 Show a feed of this user's microposts on the home page too.
Expose the user's feed from the model and use that in the
StaticPagesController.

The home page, when logged in, now has a section for the feed via a
partial which in turn displays each micropost using the existing
partial used in the profile page.

Also make sure that the feed is properly displayed even if the
micropost submit fail (which renders the home page directly from the
micropost controller as opposed to redirecting it to the home page).