14.2.2 Add stats and follow/unfollow buttons.
The stats are both on the home and profile pages.

Follow/Unfollow buttons are only on the profile page of a user and
show up depending on whether the current user is not currently
following or is already following the profile user. This is not
displayed if the viewing user is not logged in.

These are display only right now as the relationship controller is not
yet hooked up.

Add some basic test for the presence of these items.