14.3.2 User's feed now has posts from followed users.
It is a mix of the user's own posts and the posts of followed
users. This is done via a sql query for the feed.

Since the feed has microposts of other users, these won't have the
delete link. So update the feed test accordingly.

Also fix the error in relationship fixture which was causing the feed
test added in the previous commit to fail. (The last relationship was
overwriting a previous one.)